alternative in chapter 5 and other comments

+++ dectrees/python/drawtree_qt5.py
@@ -46,7 +46,7 @@ def draw(p, t, x, y):
         mid, xx = draw(p, t.branches[b], xx, y+70)
         p.drawText(mid-3, y+68, str(b))
         anchors.append(mid)
-    newMid = (x+xx)/2
+    newMid = int((x+xx)/2) # avoid the error of "unexpected type 'float'" in drawText()
     p.drawText(newMid-7, y+15, t.attribute.name)
     p.drawEllipse(newMid-15, y, 30, 20)
     for m in anchors:
